requirements.txt
flask==2.0.1
pillow>=9.0.0  
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.9.0.80
requests==2.26.0
python-dotenv==0.19.0
flask-cors==3.0.10
werkzeug==2.0.3